+++ src/main/java/com/example/demo/repositories/FatherRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FatherRepository extends JpaRepository<Father,Integer> {
+public interface FatherRepository extends JpaRepository<Father,Long> {
 }
