+++ src/main/java/com/example/demo/services/implementations/FatherService.java
@@ -2,7 +2,7 @@ package com.example.demo.services.implementations;
 
 import com.example.demo.exceptions.exceptions.DataNotFoundException;
 import com.example.demo.models.DTOs.FatherDTO;
-import com.example.demo.models.DTOs.LightDTO;
+import com.example.demo.models.dtos.LightDTO;
 import com.example.demo.models.entities.Father;
 import com.example.demo.models.mappers.FatherMapper;
 import com.example.demo.repositories.FatherRepository;
@@ -17,7 +17,7 @@ public class FatherService {
     private FatherRepository fatherRepository;
     @Autowired
     private FatherMapper fatherMapper;
-    public Integer addFather(FatherDTO fatherDTO) {
+    public Long addFather(FatherDTO fatherDTO) {
         Father father = new Father();
         fatherRepository.save(fatherMapper.fatherDtoToFather(fatherDTO, father));
         return father.getId();
@@ -27,7 +27,7 @@ public class FatherService {
         return fatherMapper.fathersToLightDtos(fatherRepository.findAll());
     }
 
-    public Father getById(Integer fatherId) {
+    public Father getById(Long fatherId) {
         Optional.ofNullable(fatherId).orElseThrow(() -> new DataNotFoundException("the father id is null"));
         return fatherRepository.findById(fatherId).orElseThrow(
                 ()-> new DataNotFoundException("the required father is not present"));
